[Improvement] Support to config byte sizes by readable string (e.g., kb/mb/gb) (#26)
### What changes were proposed in this pull request?
Support to config byte sizes by readable string (e.g., kb/mb/gb).

### Why are the changes needed?
Use number (e.g., 100000) is unreadable and easy to make mistakes.

### Does this PR introduce _any_ user-facing change?
Yes.
In Java code: shuffleServerConf.getSizeAsBytes(ShuffleServerConf.SERVER_BUFFER_CAPACITY)
In conf file: rss.server.buffer.capacity 12kb

### How was this patch tested?
see com.tencent.rss.server.ShuffleServerConfTest

+++ common/src/main/java/com/tencent/rss/common/rpc/GrpcServer.java
@@ -44,7 +44,7 @@ public class GrpcServer implements ServerInterface {
 
   public GrpcServer(RssBaseConf conf, BindableService service, GRPCMetrics grpcMetrics) {
     this.port = conf.getInteger(RssBaseConf.RPC_SERVER_PORT);
-    int maxInboundMessageSize = conf.getInteger(RssBaseConf.RPC_MESSAGE_MAX_SIZE);
+    long maxInboundMessageSize = conf.getLong(RssBaseConf.RPC_MESSAGE_MAX_SIZE);
     int rpcExecutorSize = conf.getInteger(RssBaseConf.RPC_EXECUTOR_SIZE);
     ExecutorService pool = new ThreadPoolExecutor(
         rpcExecutorSize,
@@ -63,14 +63,14 @@ public class GrpcServer implements ServerInterface {
           .forPort(port)
           .addService(ServerInterceptors.intercept(service, monitoringInterceptor))
           .executor(pool)
-          .maxInboundMessageSize(maxInboundMessageSize)
+          .maxInboundMessageSize((int)maxInboundMessageSize)
           .build();
     } else {
       this.server = ServerBuilder
           .forPort(port)
           .addService(service)
           .executor(pool)
-          .maxInboundMessageSize(maxInboundMessageSize)
+          .maxInboundMessageSize((int)maxInboundMessageSize)
           .build();
     }
   }
